requirements.txt
streamlit
pypdf
langchain
langchain-community
langchain-aws
awscli
faiss-cpu
boto3
python-dotenv